Trigger a nasty error with Webpack 2
React.createElement: type should not be null, undefined, boolean, or
number. It should be a string (for DOM elements) or a ReactClass (for
composite components). Check the render method of `AppContainer`

invariant.js:38 Uncaught Invariant Violation: Element type is invalid:
expected a string (for built-in components) or a class/function (for
composite components) but got: object. Check the render method of
`AppContainer`.